Added functionality for grouping decorator pattern elements and updated DecoratorVerifier
Added functionality for grouping classes/interfaces that are part of the same pattern instance.

Updated DecoratorVerifier to use IPatternGroupVerifier.
Updated DecoratorVerifier to not use instance variables (otherwise the variables would need to be reset between calls to verify for a clean slate).

The stuff in main is a test case that needs to be moved to its own file.